PIM-5908: Migrate the enhanced csv attribute export
