+++ DependencyInjection/PimEnhancedConnectorExtension.php
@@ -22,6 +22,8 @@ class PimEnhancedConnectorExtension extends Extension
     public function load(array $configs, ContainerBuilder $container)
     {
         $loader = new YamlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
+        $loader->load('jobs.yml');
         $loader->load('processors.yml');
+        $loader->load('steps.yml');
     }
 }
