+++ Resources/config/jobs.yml
@@ -1,0 +1,39 @@
+parameters:
+    pim_enhanced_connector.connector_name: 'Enhanced Connector'
+    pim_enhanced_connector.job_name.csv_attribute_export: 'enhanced_csv_attribute_export'
+
+services:
+    pim_enhanced_connector.job.csv_attribute_export:
+        class: '%pim_connector.job.simple_job.class%'
+        arguments:
+            - '%pim_enhanced_connector.job_name.csv_attribute_export%'
+            - '@event_dispatcher'
+            - '@akeneo_batch.job_repository'
+            -
+                - '@pim_enhanced_connector.step.csv_attribute.export'
+        tags:
+            - { name: akeneo_batch.job, connector: '%pim_enhanced_connector.connector_name%', type: '%pim_connector.job.export_type%' }
+
+    pim_enhanced_connector.job.job_parameters.default_values_provider.enhanced_csv_export:
+        class: '%pim_connector.job.job_parameters.default_values_provider.simple_csv_export.class%'
+        arguments:
+            -
+                - 'enhanced_csv_attribute_export'
+        tags:
+            - { name: akeneo_batch.job.job_parameters.default_values_provider }
+
+    pim_enhanced_connector.job.job_parameters.constraint_collection_provider.enhanced_csv_export:
+        class: '%pim_connector.job.job_parameters.constraint_collection_provider.simple_csv_export.class%'
+        arguments:
+            -
+                - 'enhanced_csv_attribute_export'
+        tags:
+            - { name: akeneo_batch.job.job_parameters.constraint_collection_provider }
+
+    pim_enhanced_connector.job_parameters.form_configuration_provider.enhanced_csv_export:
+        class: '%pim_import_export.job_parameters.form_configuration_provider.simple_csv_export.class%'
+        arguments:
+            -
+                - 'enhanced_csv_attribute_export'
+        tags:
+            - { name: pim_import_export.job_parameters.form_configuration_provider }
